requirements.txt
pandas==2.2.2
protobuf==5.26.1
python-dotenv==1.0.1
streamlit==1.33.0
google-generativeai==0.5.0

